update all package.json deps migrate from enzyme to react-testing library change specs folder change imports/exports (seperate .types files) rewrite to functional components with hooks where possible add useFormApi hook to the exports
