+++ packages/lib/__tests__/components/InputAdapter/index.ts
