+++ packages/lib/__tests__/specs/render.tsx
@@ -1,12 +1,16 @@
-import * as React from 'react';
-import {mount} from 'enzyme';
-import {TestForm} from '../test/components/TestForm';
-import {createTestFormDriver} from '../test/components/TestForm.driver';
-import {InputAdapter} from '../test/components/InputAdapter';
-import {Field} from '../src';
+import {cleanup, render} from '@testing-library/react';
+import React from 'react';
+import {TestForm} from '../components/TestForm';
+import {createTestFormDriver} from '../components/TestForm.driver';
+import {InputAdapter} from '../components/InputAdapter';
+import {Field} from '../../src';
 
 test('Render', () => {
-  const wrapper = mount(
+  afterEach(() => {
+    return cleanup();
+  });
+
+  const wrapper = render(
     <TestForm>
       <Field name={TestForm.FIELD_ONE_NAME} adapter={InputAdapter} />
       <Field name={TestForm.FIELD_TWO_NAME}>
@@ -15,7 +19,7 @@ test('Render', () => {
         }}
       </Field>
     </TestForm>,
-  );
+  ).container;
   const formDriver = createTestFormDriver({wrapper});
 
   expect(formDriver.get.serialized()).toMatchSnapshot();
