+++ packages/lib/__tests__/specs/useFormAPI.tsx
@@ -1,28 +1,31 @@
-import * as React from 'react';
-import {mount} from 'enzyme';
-import {TestForm} from '../test/components/TestForm';
-import {Field} from '../src/Field/Field';
-import {InputAdapter} from '../test/components/InputAdapter';
-import {FieldValueDisplay} from '../test/components/FieldValueDisplay/FieldValueDisplay';
-import {createInputAdapterDriver} from '../test/components/InputAdapter/InputAdapter.driver';
-import {createFieldValueDisplayDriver} from '../test/components/FieldValueDisplay/FieldValueDisplay.driver';
+import {cleanup, render} from '@testing-library/react';
+import React from 'react';
+import {TestForm} from '../components/TestForm';
+import {Field} from '../../src/Field';
+import {InputAdapter} from '../components/InputAdapter';
+import {FieldValueDisplayWithHook, createFieldValueDisplayDriver} from '../components/FieldValueDisplay';
+import {createInputAdapterDriver} from '../components/InputAdapter/InputAdapter.driver';
 
-describe('inhectFormApi', () => {
+describe('useFormController', () => {
+  afterEach(() => {
+    return cleanup();
+  });
+
   const TEST_INITIAL_VALUE = 'testtest';
   const NEW_VALUE = 'new-value';
   const INNER_FORM_COMPONENT_DATA_HOOK = 'inner-data-hook';
 
-  it('should pass formApi with `injectFormApi` decorator', () => {
-    const wrapper = mount(
+  it('should pass formApi with `useFormApi` hook', () => {
+    const wrapper = render(
       <TestForm
         initialValues={{
           [TestForm.FIELD_ONE_NAME]: TEST_INITIAL_VALUE,
         }}
       >
         <Field name={TestForm.FIELD_ONE_NAME} adapter={InputAdapter} />
-        <FieldValueDisplay dataHook={INNER_FORM_COMPONENT_DATA_HOOK} displayedFieldName={TestForm.FIELD_ONE_NAME} />
+        <FieldValueDisplayWithHook dataHook={INNER_FORM_COMPONENT_DATA_HOOK} displayedFieldName={TestForm.FIELD_ONE_NAME} />
       </TestForm>,
-    );
+    ).container;
 
     const fieldDriver = createInputAdapterDriver({wrapper, dataHook: TestForm.FIELD_ONE_NAME});
     const FieldValueDisplayDriver = createFieldValueDisplayDriver({
@@ -31,6 +34,7 @@ describe('inhectFormApi', () => {
     });
 
     expect(FieldValueDisplayDriver.get.text()).toEqual(TEST_INITIAL_VALUE);
+
     fieldDriver.when.change(NEW_VALUE);
 
     expect(FieldValueDisplayDriver.get.text()).toEqual(NEW_VALUE);
