+++ packages/lib/package.json
@@ -1,7 +1,8 @@
 {
   "name": "formagus",
-  "version": "0.0.31",
+  "version": "1.0.0",
   "license": "MIT",
+  "sideEffects": false,
   "private": false,
   "authors": [
     {
@@ -10,15 +11,15 @@
     }
   ],
   "engines": {
-    "node": ">=8"
+    "node": ">=12"
   },
   "scripts": {
     "build": "webpack --config webpack.config.js",
     "test": "jest",
-    "test-unit": "jest",
-    "test-watch": "jest --watch --detectOpenHandles",
-    "test-coverage": "jest --coverage",
-    "posttest": "npm run test-coverage",
+    "test:unit": "jest",
+    "test:watch": "jest --watch --detectOpenHandles",
+    "test:coverage": "jest --coverage",
+    "posttest": "npm run test:coverage",
     "precommit": "lint-staged",
     "postcommit": "git update-index --again",
     "prepush": "npm run test",
@@ -41,44 +42,40 @@
     "/dist/index.js"
   ],
   "devDependencies": {
-    "@types/enzyme": "^3.1.13",
-    "@types/enzyme-adapter-react-16": "^1.0.3",
-    "@types/flat": "^0.0.28",
-    "@types/jest": "^23.3.9",
-    "@types/lodash-es": "^4.17.1",
-    "@types/react": "^16.7.3",
-    "@types/react-dom": "^16.0.7",
-    "commitizen": "^3.0.4",
-    "cz-conventional-changelog": "^2.1.0",
-    "enzyme": "^3.4.1",
-    "enzyme-adapter-react-16": "^1.7.0",
-    "enzyme-to-json": "^3.3.4",
-    "husky": "^1.1.2",
-    "jest": "^23.5.0",
-    "jest-cli": "^23.6.0",
-    "jest-watch-typeahead": "^0.2.0",
-    "lint-staged": "^8.0.4",
-    "mobx": "^5.5.2",
-    "mobx-react": "^5.2.5",
-    "prettier": "^1.15.2",
-    "react": "^16.6.1",
-    "react-dom": "^16.6.1",
-    "react-test-renderer": "^16.6.1",
+    "@testing-library/react": "^10.4.8",
+    "@types/flat": "^5.0.1",
+    "@types/hoist-non-react-statics": "^3.3.1",
+    "@types/jest": "^26.0.9",
+    "@types/react": "^16.9.46",
+    "@types/react-dom": "^16.9.8",
+    "commitizen": "^4.1.2",
+    "cz-conventional-changelog": "^3.2.0",
+    "husky": "^4.2.5",
+    "jest": "^26.3.0",
+    "jest-cli": "^26.3.0",
+    "jest-watch-typeahead": "^0.6.0",
+    "lint-staged": "^10.2.11",
+    "mobx": "^5.15.5",
+    "mobx-react": "^6.2.5",
+    "prettier": "^2.0.5",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "trier-promise": "^1.0.1",
-    "ts-jest": "^23.1.3",
-    "ts-loader": "^5.3.0",
-    "tslib": "^1.9.3",
-    "tslint": "^5.11.0",
-    "tslint-config-prettier": "^1.14.0",
+    "ts-jest": "^26.2.0",
+    "ts-loader": "^8.0.2",
+    "tslib": "^2.0.1",
+    "tslint": "^6.1.3",
+    "tslint-config-prettier": "^1.18.0",
     "tslint-config-wix": "^1.0.31",
-    "typescript": "^3.1.4",
-    "webpack": "^4.16.5",
-    "webpack-cli": "^3.1.0"
+    "typescript": "^3.9.7",
+    "uglifyjs-webpack-plugin": "^2.2.0",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12"
   },
   "dependencies": {
     "flat": "^4.1.0",
-    "hoist-non-react-statics": "^3.3.0",
-    "lodash": "^4.17.8"
+    "hoist-non-react-statics": "^3.3.2",
+    "lodash": "^4.17.19"
   },
   "peerDependencies": {
     "mobx": "*",
