+++ packages/lib/src/Field/index.ts
@@ -1,1 +1,3 @@
+export * from './FieldClass';
 export * from './Field';
+export * from './Field.types';
