+++ packages/lib/src/Form/Form.tsx
@@ -1,18 +1,12 @@
-import * as React from 'react';
-import {observer, Provider} from 'mobx-react';
-import {FormController, FormControllerOptions} from '../FormController';
-import {FormPart, FormPartProps} from '../FormPart';
+import {Observer, Provider} from 'mobx-react';
+import React, {useState, FC} from 'react';
+import {FormController} from '../FormController';
+import {FormPart} from '../FormPart';
+import {FormControllerContext} from './FormControllerContext';
+import type {FormProps} from './Form.types';
 
-export interface FormProps extends FormControllerOptions, FormPartProps {}
-
-@observer
-export class Form extends React.Component<FormProps> {
-  controller: FormController;
-
-  constructor(props: any) {
-    super(props);
-
-    //"children" will be another prop, so take it to the account as well
+export const Form: FC<FormProps> = (props) => {
+  const [controller] = useState(() => {
     if (props.controller && Object.keys(props).length > 2) {
       throw new Error(
         'Form should have either "controller" prop with configured Controller instance or no ' +
@@ -21,15 +15,19 @@ export class Form extends React.Component<FormProps> {
     }
     //controller can be injected by prop and created in any place,
     //or be created on the flight with passed configuration through props
-    this.controller = props.controller || new FormController(props);
-  }
+    return props.controller || new FormController(props);
+  });
 
   //creates the provider and sets the controller, which will control all the form state
-  render() {
-    return (
-      <Provider controller={this.controller}>
-        <FormPart>{this.props.children}</FormPart>
-      </Provider>
-    );
-  }
-}
+  return (
+    <Observer>
+      {() => (
+        <Provider controller={controller}>
+          <FormControllerContext.Provider value={controller}>
+            <FormPart>{props.children}</FormPart>
+          </FormControllerContext.Provider>
+        </Provider>
+      )}
+    </Observer>
+  );
+};
