+++ packages/lib/src/Form/index.ts
@@ -1,1 +1,3 @@
 export * from './Form';
+export * from './Form.types';
+export * from './useFormController';
