+++ packages/lib/src/Form/useFormController.ts
@@ -1,0 +1,6 @@
+import {useContext} from 'react';
+import {FormControllerContext} from './FormControllerContext';
+
+export const useFormController = () => {
+  return useContext(FormControllerContext);
+};
