+++ packages/lib/src/FormController/index.ts
@@ -1,1 +1,2 @@
 export * from './FormController';
+export * from './FormController.types';
