+++ packages/lib/src/injectFormApi/injectFormApi.tsx
@@ -1,19 +1,15 @@
-import * as React from 'react';
+import React from 'react';
 import {inject, observer} from 'mobx-react';
-import {FormController} from '../FormController/FormController';
-import * as hoistNonReactStatics from 'hoist-non-react-statics';
-
-interface InjectFormApiWrapperProps {
-  controller?: FormController;
-}
+import hoistNonReactStatics from 'hoist-non-react-statics';
+import type {FormController} from '../FormController';
 
 function ComponentWithInjectedFormApi<C extends React.ComponentClass>(Component: C): C {
   @inject('controller')
   @observer
-  class InjectFormApiWrapper extends React.Component<InjectFormApiWrapperProps> {
+  class InjectFormApiWrapper extends React.Component<{controller?: FormController}> {
     render() {
       const {controller, ...props} = this.props;
-      return <Component {...props as any} formApi={controller!.API} />;
+      return <Component {...(props as any)} formApi={controller!.API} />;
     }
   }
 
@@ -22,4 +18,5 @@ function ComponentWithInjectedFormApi<C extends React.ComponentClass>(Component:
   return InjectFormApiWrapper as C;
 }
 
+/** @deprecated */
 export const injectFormApi = <C extends React.ComponentClass>(Component: C) => ComponentWithInjectedFormApi(Component);
