+++ packages/lib/src/useFormApi/useFormApi.tsx
@@ -1,0 +1,17 @@
+import {autorun} from 'mobx';
+import {useEffect, useState} from 'react';
+import {useFormController} from '../Form';
+import type {FormAPI} from '../FormController';
+
+export const useFormApi = () => {
+  const controller = useFormController();
+  const [api, setApi] = useState<FormAPI>(() => controller.API);
+
+  useEffect(() => {
+    return autorun(() => {
+      setApi(controller.API);
+    });
+  }, []);
+
+  return api;
+};
