--- packages/lib/test/components/FieldValueDisplay/index.ts
@@ -1,1 +1,0 @@
-export * from './FieldValueDisplay';
