--- packages/lib/test/jest-setup.ts
@@ -1,6 +1,0 @@
-import {configure} from 'enzyme';
-import * as Adapter from 'enzyme-adapter-react-16';
-
-configure({
-  adapter: new Adapter(),
-});
