--- packages/lib/typings/externals.d.ts
@@ -1,1 +1,0 @@
-declare module 'hoist-non-react-statics';
