+++ packages/lib/webpack.config.js
@@ -13,28 +13,17 @@ module.exports = {
       },
     ],
   },
-  plugins: [
-    new UglifyJsPlugin({
+  optimization: {
+    minimize: true,
+    minimizer: [new UglifyJsPlugin({
       uglifyOptions: {
+        warnings: false,
         mangle: {
           toplevel: true,
         },
-        beautify: false,
-        compress: {
-          warnings: false,
-          conditionals: true,
-          unused: true,
-          comparisons: true,
-          sequences: true,
-          dead_code: true,
-          evaluate: true,
-          if_return: true,
-          join_vars: true,
-          negate_iife: false,
-        },
       },
-    }),
-  ],
+    })],
+  },
   mode: 'production',
   externals: {
     react: 'react',
